+++ include/freerdp/crypto/certificate.h
@@ -2,7 +2,8 @@
  * FreeRDP: A Remote Desktop Protocol Implementation
  * Certificate Handling
  *
- * Copyright 2011-2012 Marc-Andre Moreau <marcandre.moreau@gmail.com>
+ * Copyright 2023 Armin Novak <anovak@thincast.com>
+ * Copyright 2023 Thincast Technologies GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,58 +21,60 @@
 #ifndef FREERDP_CRYPTO_CERTIFICATE_H
 #define FREERDP_CRYPTO_CERTIFICATE_H
 
-typedef struct rdp_certificate_data rdpCertificateData;
-typedef struct rdp_certificate_store rdpCertificateStore;
-
-#include <freerdp/crypto/ber.h>
-#include <freerdp/crypto/crypto.h>
+#include <winpr/crypto.h>
 
 #include <freerdp/api.h>
-#include <freerdp/settings.h>
-
-#include <winpr/print.h>
-#include <winpr/stream.h>
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-	FREERDP_API rdpCertificateData* certificate_data_new(const char* hostname, UINT16 port);
-	FREERDP_API void certificate_data_free(rdpCertificateData* certificate_data);
+	typedef struct rdp_certificate rdpCertificate;
 
-	FREERDP_API const char* certificate_data_get_host(const rdpCertificateData* cert);
-	FREERDP_API UINT16 certificate_data_get_port(const rdpCertificateData* cert);
+	FREERDP_API rdpCertificate* freerdp_certificate_new(void);
+	FREERDP_API rdpCertificate* freerdp_certificate_new_from_file(const char* file);
+	FREERDP_API rdpCertificate* freerdp_certificate_new_from_pem(const char* pem);
+	FREERDP_API rdpCertificate* freerdp_certificate_new_from_der(const BYTE* data, size_t length);
 
-	FREERDP_API BOOL certificate_data_set_pem(rdpCertificateData* cert, const char* pem);
-	FREERDP_API BOOL certificate_data_set_subject(rdpCertificateData* cert, const char* subject);
-	FREERDP_API BOOL certificate_data_set_issuer(rdpCertificateData* cert, const char* issuer);
-	FREERDP_API BOOL certificate_data_set_fingerprint(rdpCertificateData* cert,
-	                                                  const char* fingerprint);
-	FREERDP_API const char* certificate_data_get_pem(const rdpCertificateData* cert);
-	FREERDP_API const char* certificate_data_get_subject(const rdpCertificateData* cert);
-	FREERDP_API const char* certificate_data_get_issuer(const rdpCertificateData* cert);
-	FREERDP_API const char* certificate_data_get_fingerprint(const rdpCertificateData* cert);
+	FREERDP_API void freerdp_certificate_free(rdpCertificate* certificate);
 
-	FREERDP_API rdpCertificateStore* certificate_store_new(const rdpSettings* settings);
-	FREERDP_API void certificate_store_free(rdpCertificateStore* certificate_store);
+	FREERDP_API char* freerdp_certificate_get_hash(const rdpCertificate* certificate,
+	                                               const char* hash, size_t* plength);
 
-	FREERDP_API int certificate_store_contains_data(rdpCertificateStore* certificate_store,
-	                                                const rdpCertificateData* certificate_data);
-	FREERDP_API rdpCertificateData*
-	certificate_store_load_data(rdpCertificateStore* certificate_store, const char* host,
-	                            UINT16 port);
-	FREERDP_API BOOL certificate_store_save_data(rdpCertificateStore* certificate_store,
-	                                             const rdpCertificateData* certificate_data);
-	FREERDP_API BOOL certificate_store_remove_data(rdpCertificateStore* certificate_store,
-	                                               const rdpCertificateData* certificate_data);
+	FREERDP_API char* freerdp_certificate_get_fingerprint_by_hash(const rdpCertificate* certificate,
+	                                                              const char* hash);
+	FREERDP_API char*
+	freerdp_certificate_get_fingerprint_by_hash_ex(const rdpCertificate* certificate,
+	                                               const char* hash, BOOL separator);
+	FREERDP_API char* freerdp_certificate_get_fingerprint(const rdpCertificate* certificate);
+	FREERDP_API char* freerdp_certificate_get_pem(const rdpCertificate* certificate,
+	                                              size_t* pLength);
+	FREERDP_API BYTE* freerdp_certificate_get_der(const rdpCertificate* certificate,
+	                                              size_t* pLength);
 
-	FREERDP_API const char*
-	certificate_store_get_hosts_file(const rdpCertificateStore* certificate_store);
-	FREERDP_API const char*
-	certificate_store_get_certs_path(const rdpCertificateStore* certificate_store);
-	FREERDP_API const char*
-	certificate_store_get_hosts_path(const rdpCertificateStore* certificate_store);
+	FREERDP_API char* freerdp_certificate_get_subject(const rdpCertificate* certificate);
+	FREERDP_API char* freerdp_certificate_get_issuer(const rdpCertificate* certificate);
+
+	FREERDP_API char* freerdp_certificate_get_upn(const rdpCertificate* certificate);
+	FREERDP_API char* freerdp_certificate_get_email(const rdpCertificate* certificate);
+
+	FREERDP_API WINPR_MD_TYPE freerdp_certificate_get_signature_alg(const rdpCertificate* cert);
+
+	FREERDP_API char* freerdp_certificate_get_common_name(const rdpCertificate* cert,
+	                                                      size_t* plength);
+	FREERDP_API char** freerdp_certificate_get_dns_names(const rdpCertificate* cert, size_t* pcount,
+	                                                     size_t** pplengths);
+	FREERDP_API void freerdp_certificate_free_dns_names(size_t count, size_t* lengths,
+	                                                    char** names);
+
+	FREERDP_API BOOL freerdp_certificate_check_eku(const rdpCertificate* certificate, int nid);
+
+	FREERDP_API BOOL freerdp_certificate_get_public_key(const rdpCertificate* cert,
+	                                                    BYTE** PublicKey, DWORD* PublicKeyLength);
+
+	FREERDP_API BOOL freerdp_certificate_verify(const rdpCertificate* cert,
+	                                            const char* certificate_store_path);
 
 #ifdef __cplusplus
 }
