+++ libfreerdp/core/settings.c
@@ -18,9 +18,7 @@
  */
 
 #include <freerdp/config.h>
-
-#include "certificate.h"
-#include "capabilities.h"
+#include <freerdp/crypto/certificate.h>
 
 #include <ctype.h>
 
@@ -34,6 +32,9 @@
 #include <freerdp/build-config.h>
 #include <ctype.h>
 
+#include "../crypto/certificate.h"
+#include "../crypto/privatekey.h"
+#include "capabilities.h"
 #include "settings.h"
 
 #define TAG FREERDP_TAG("settings")
@@ -336,7 +337,6 @@ BOOL freerdp_capability_buffer_allocate(rdpSettings* settings, UINT32 count)
 
 rdpSettings* freerdp_settings_new(DWORD flags)
 {
-	size_t x;
 	char* base;
 	char* issuers[] = { "FreeRDP", "FreeRDP-licenser" };
 	rdpSettings* settings = (rdpSettings*)calloc(1, sizeof(rdpSettings));
@@ -432,7 +432,6 @@ rdpSettings* freerdp_settings_new(DWORD flags)
 	    !freerdp_settings_set_bool(settings, FreeRDP_DisableCredentialsDelegation, FALSE) ||
 	    !freerdp_settings_set_uint32(settings, FreeRDP_AuthenticationLevel, 2) ||
 	    !freerdp_settings_set_uint32(settings, FreeRDP_ChannelCount, 0) ||
-	    !freerdp_settings_set_bool(settings, FreeRDP_CertificateUseKnownHosts, TRUE) ||
 	    !freerdp_settings_set_bool(settings, FreeRDP_CertificateCallbackPreferPEM, FALSE) ||
 	    !freerdp_settings_set_uint32(settings, FreeRDP_KeySpec, AT_KEYEXCHANGE))
 		goto out_fail;
@@ -554,7 +553,7 @@ rdpSettings* freerdp_settings_new(DWORD flags)
 	if (!settings->FragCache)
 		goto out_fail;
 
-	for (x = 0; x < 10; x++)
+	for (size_t x = 0; x < 10; x++)
 	{
 		GLYPH_CACHE_DEFINITION cache = { 0 };
 		cache.cacheEntries = 254;
@@ -827,7 +826,9 @@ static BOOL freerdp_settings_int_buffer_copy(rdpSettings* _settings, const rdpSe
 	}
 	if (settings->RdpServerCertificate)
 	{
-		rdpCertificate* cert = certificate_clone(settings->RdpServerCertificate);
+		rdpCertificate* cert = freerdp_certificate_clone(settings->RdpServerCertificate);
+		if (!cert)
+			goto out_fail;
 		if (!freerdp_settings_set_pointer_len(_settings, FreeRDP_RdpServerCertificate, cert, 1))
 			goto out_fail;
 	}
@@ -839,7 +840,9 @@ static BOOL freerdp_settings_int_buffer_copy(rdpSettings* _settings, const rdpSe
 
 	if (settings->RdpServerRsaKey)
 	{
-		rdpRsaKey* key = key_clone(settings->RdpServerRsaKey);
+		rdpRsaKey* key = freerdp_key_clone(settings->RdpServerRsaKey);
+		if (!key)
+			goto out_fail;
 		if (!freerdp_settings_set_pointer_len(_settings, FreeRDP_RdpServerRsaKey, key, 1))
 			goto out_fail;
 	}
