+++ libfreerdp/core/test/settings_property_lists.h
@@ -23,7 +23,6 @@ static const size_t bool_list_indices[] = {
 	FreeRDP_BitmapCacheV3Enabled,
 	FreeRDP_BitmapCompressionDisabled,
 	FreeRDP_CertificateCallbackPreferPEM,
-	FreeRDP_CertificateUseKnownHosts,
 	FreeRDP_CompressionEnabled,
 	FreeRDP_ConsoleSession,
 	FreeRDP_CredentialsFromStdin,
