+++ libfreerdp/crypto/CMakeLists.txt
@@ -24,7 +24,16 @@ freerdp_module_add(
 	ber.c
 	per.c
 	base64.c
-	certificate.c
+    x509_utils.c
+    x509_utils.h
+    cert_common.h
+    cert_common.c
+    privatekey.c
+    privatekey.h
+    certificate.c
+    certificate.h
+	certificate_data.c
+	certificate_store.c
 	crypto.c
 	tls.c
 	tls.h
