+++ libfreerdp/crypto/certificate.c
@@ -3,7 +3,9 @@
  * Certificate Handling
  *
  * Copyright 2011 Jiten Pathy
- * Copyright 2011-2012 Marc-Andre Moreau <marcandre.moreau@gmail.com>
+ * Copyright 2011 Marc-Andre Moreau <marcandre.moreau@gmail.com>
+ * Copyright 2015 Thincast Technologies GmbH
+ * Copyright 2015 DI (FH) Martin Haimberger <martin.haimberger@thincast.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,1156 +22,1473 @@
 
 #include <freerdp/config.h>
 
-#include <winpr/assert.h>
 #include <errno.h>
 #include <stdio.h>
 #include <string.h>
-#include <ctype.h>
 
-#include <winpr/crypto.h>
+#include <winpr/assert.h>
+#include <winpr/wtypes.h>
 #include <winpr/crt.h>
 #include <winpr/file.h>
-#include <winpr/path.h>
+#include <winpr/print.h>
+#include <winpr/crypto.h>
 
+#include <freerdp/crypto/certificate.h>
+
+#include <openssl/err.h>
 #include <openssl/pem.h>
 #include <openssl/rsa.h>
 
-#include <freerdp/settings.h>
+#include "certificate.h"
+#include "cert_common.h"
+#include "crypto.h"
 
-#include <freerdp/crypto/certificate.h>
+#include "x509_utils.h"
+#include "privatekey.h"
+#include "opensslcompat.h"
 
-struct rdp_certificate_store
+#define TAG FREERDP_TAG("core")
+
+#define CERTIFICATE_TAG FREERDP_TAG("core.certificate")
+#ifdef WITH_DEBUG_CERTIFICATE
+#define DEBUG_CERTIFICATE(...) WLog_DBG(TAG, __VA_ARGS__)
+#else
+#define DEBUG_CERTIFICATE(...) \
+	do                         \
+	{                          \
+	} while (0)
+#endif
+
+#define TSSK_KEY_LENGTH 64
+
+struct rdp_CertBlob
 {
-	char* file;
-	char* certs_path;
-	char* server_path;
-	const rdpSettings* settings;
+	UINT32 length;
+	BYTE* data;
+};
+typedef struct rdp_CertBlob rdpCertBlob;
+
+struct rdp_X509CertChain
+{
+	UINT32 count;
+	rdpCertBlob* array;
+};
+typedef struct rdp_X509CertChain rdpX509CertChain;
+
+struct rdp_certificate
+{
+	X509* x509;
+	STACK_OF(X509) * chain;
+
+	BOOL isRSA;
+
+	rdpCertInfo cert_info;
+	rdpX509CertChain x509_cert_chain;
 };
 
-struct rdp_certificate_data
-{
-	char* hostname;
-	UINT16 port;
-	char* subject;
-	char* issuer;
-	char* fingerprint;
-	char* pem;
+/**
+ *
+ * X.509 Certificate Structure
+ *
+ * Certificate ::= SEQUENCE
+ * {
+ * 	tbsCertificate			TBSCertificate,
+ * 	signatureAlgorithm		AlgorithmIdentifier,
+ * 	signatureValue			BIT_STRING
+ * }
+ *
+ * TBSCertificate ::= SEQUENCE
+ * {
+ * 	version			[0]	EXPLICIT Version DEFAULT v1,
+ * 	serialNumber			CertificateSerialNumber,
+ * 	signature			AlgorithmIdentifier,
+ * 	issuer				Name,
+ * 	validity			Validity,
+ * 	subject				Name,
+ * 	subjectPublicKeyInfo		SubjectPublicKeyInfo,
+ * 	issuerUniqueID		[1]	IMPLICIT UniqueIdentifier OPTIONAL,
+ * 	subjectUniqueId		[2]	IMPLICIT UniqueIdentifier OPTIONAL,
+ * 	extensions		[3]	EXPLICIT Extensions OPTIONAL
+ * }
+ *
+ * Version ::= INTEGER { v1(0), v2(1), v3(2) }
+ *
+ * CertificateSerialNumber ::= INTEGER
+ *
+ * AlgorithmIdentifier ::= SEQUENCE
+ * {
+ * 	algorithm			OBJECT_IDENTIFIER,
+ * 	parameters			ANY DEFINED BY algorithm OPTIONAL
+ * }
+ *
+ * Name ::= CHOICE { RDNSequence }
+ *
+ * RDNSequence ::= SEQUENCE OF RelativeDistinguishedName
+ *
+ * RelativeDistinguishedName ::= SET OF AttributeTypeAndValue
+ *
+ * AttributeTypeAndValue ::= SEQUENCE
+ * {
+ * 	type				AttributeType,
+ * 	value				AttributeValue
+ * }
+ *
+ * AttributeType ::= OBJECT_IDENTIFIER
+ *
+ * AttributeValue ::= ANY DEFINED BY AttributeType
+ *
+ * Validity ::= SEQUENCE
+ * {
+ * 	notBefore			Time,
+ * 	notAfter			Time
+ * }
+ *
+ * Time ::= CHOICE
+ * {
+ * 	utcTime				UTCTime,
+ * 	generalTime			GeneralizedTime
+ * }
+ *
+ * UniqueIdentifier ::= BIT_STRING
+ *
+ * SubjectPublicKeyInfo ::= SEQUENCE
+ * {
+ * 	algorithm			AlgorithmIdentifier,
+ * 	subjectPublicKey		BIT_STRING
+ * }
+ *
+ * RSAPublicKey ::= SEQUENCE
+ * {
+ * 	modulus				INTEGER
+ * 	publicExponent			INTEGER
+ * }
+ *
+ * Extensions ::= SEQUENCE SIZE (1..MAX) OF Extension
+ *
+ * Extension ::= SEQUENCE
+ * {
+ * 	extnID				OBJECT_IDENTIFIER
+ * 	critical			BOOLEAN DEFAULT FALSE,
+ * 	extnValue			OCTET_STRING
+ * }
+ *
+ */
+
+static const char rsa_magic[4] = "RSA1";
+
+static const char* certificate_read_errors[] = { "Certificate tag",
+	                                             "TBSCertificate",
+	                                             "Explicit Contextual Tag [0]",
+	                                             "version",
+	                                             "CertificateSerialNumber",
+	                                             "AlgorithmIdentifier",
+	                                             "Issuer Name",
+	                                             "Validity",
+	                                             "Subject Name",
+	                                             "SubjectPublicKeyInfo Tag",
+	                                             "subjectPublicKeyInfo::AlgorithmIdentifier",
+	                                             "subjectPublicKeyInfo::subjectPublicKey",
+	                                             "RSAPublicKey Tag",
+	                                             "modulusLength",
+	                                             "zero padding",
+	                                             "modulusLength",
+	                                             "modulus",
+	                                             "publicExponent length",
+	                                             "publicExponent" };
+
+static const BYTE initial_signature[] = {
+	0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+	0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+	0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+	0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x01
 };
 
-static const char certificate_store_dir[] = "certs";
-static const char certificate_server_dir[] = "server";
-static const char certificate_known_hosts_file[] = "known_hosts2";
+#if defined(CERT_VALIDATE_RSA)
+static const BYTE tssk_exponent[] = { 0x5b, 0x7b, 0x88, 0xc0 };
+#endif
 
-#include <freerdp/log.h>
-#include <freerdp/crypto/certificate.h>
+static void certificate_free_int(rdpCertificate* certificate);
+static BOOL cert_clone_int(rdpCertificate* dst, const rdpCertificate* src);
 
-#define TAG FREERDP_TAG("crypto")
+/* [MS-RDPBCGR] 5.3.3.2 X.509 Certificate Chains:
+ *
+ * More detail[MS-RDPELE] section 2.2.1.4.2.
+ */
+static BOOL cert_blob_copy(rdpCertBlob* dst, const rdpCertBlob* src);
+static void cert_blob_free(rdpCertBlob* blob);
+static BOOL cert_blob_write(const rdpCertBlob* blob, wStream* s);
+static BOOL cert_blob_read(rdpCertBlob* blob, wStream* s);
 
-static BOOL certificate_get_file_data(rdpCertificateStore* store, rdpCertificateData* data);
-static BOOL duplicate(char** data, const char* value);
-
-static HANDLE open_file(const char* name, DWORD dwDesiredAccess, DWORD dwShareMode,
-                        DWORD dwCreationDisposition, DWORD dwFlagsAndAttributes)
+BOOL cert_blob_read(rdpCertBlob* blob, wStream* s)
 {
-	HANDLE fp;
-	if (!name)
-		return INVALID_HANDLE_VALUE;
-	WCHAR* wfile = ConvertUtf8ToWCharAlloc(name, NULL);
-	if (!wfile)
-		return INVALID_HANDLE_VALUE;
+	UINT32 certLength = 0;
+	WINPR_ASSERT(blob);
+	cert_blob_free(blob);
 
-	fp = CreateFileW(wfile, dwDesiredAccess, 0, NULL, dwCreationDisposition, dwFlagsAndAttributes,
-	                 NULL);
-	free(wfile);
-	return fp;
-}
-static rdpCertificateData* certificate_split_line(char* line);
-static BOOL certificate_line_is_comment(const char* line, size_t length)
-{
-	while (length > 0)
-	{
-		switch (*line)
-		{
-			case ' ':
-			case '\t':
-				line++;
-				length--;
-				break;
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, 4))
+		return FALSE;
 
-			case '#':
-				return TRUE;
+	Stream_Read_UINT32(s, certLength);
 
-			default:
-				return FALSE;
-		}
-	}
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, certLength))
+		return FALSE;
+
+	DEBUG_CERTIFICATE("X.509 Certificate #%" PRIu32 ", length:%" PRIu32 "", i + 1, certLength);
+	blob->data = (BYTE*)malloc(certLength);
+
+	if (!blob->data)
+		return FALSE;
+
+	Stream_Read(s, blob->data, certLength);
+	blob->length = certLength;
 
 	return TRUE;
 }
 
-static void certificate_store_uninit(rdpCertificateStore* certificate_store)
+BOOL cert_blob_write(const rdpCertBlob* blob, wStream* s)
 {
-	if (certificate_store)
-	{
-		free(certificate_store->certs_path);
-		certificate_store->certs_path = NULL;
+	WINPR_ASSERT(blob);
 
-		free(certificate_store->file);
-		certificate_store->file = NULL;
-
-		free(certificate_store->server_path);
-		certificate_store->server_path = NULL;
-	}
-}
-
-static BOOL ensure_path_exists(const char* path)
-{
-	BOOL res = FALSE;
-	if (!path)
-		return FALSE;
-	/* Use wide character functions to allow proper unicode handling on windows */
-	WCHAR* wpath = ConvertUtf8ToWCharAlloc(path, NULL);
-
-	if (!wpath)
+	if (!Stream_EnsureRemainingCapacity(s, 4 + blob->length))
 		return FALSE;
 
-	if (!PathFileExistsW(wpath))
-	{
-		if (!PathMakePathW(wpath, 0))
-		{
-			WLog_ERR(TAG, "error creating directory '%s'", path);
-			goto fail;
-		}
-
-		WLog_INFO(TAG, "creating directory %s", path);
-	}
-	res = TRUE;
-fail:
-	free(wpath);
-	return res;
-}
-
-static BOOL certificate_store_init(rdpCertificateStore* certificate_store)
-{
-	const rdpSettings* settings;
-	const char* ConfigPath;
-
-	if (!certificate_store)
-		return FALSE;
-
-	settings = certificate_store->settings;
-	if (!settings)
-		return FALSE;
-
-	ConfigPath = freerdp_settings_get_string(settings, FreeRDP_ConfigPath);
-	if (!ConfigPath)
-		return FALSE;
-
-	WINPR_ASSERT(!certificate_store->certs_path);
-	if (!(certificate_store->certs_path =
-	          GetCombinedPath(ConfigPath, (const char*)certificate_store_dir)))
-		goto fail;
-
-	WINPR_ASSERT(!certificate_store->server_path);
-	certificate_store->server_path =
-	    GetCombinedPath(ConfigPath, (const char*)certificate_server_dir);
-	if (!certificate_store->server_path)
-		goto fail;
-
-	WINPR_ASSERT(!certificate_store->file);
-	if (!(certificate_store->file =
-	          GetCombinedPath(ConfigPath, (const char*)certificate_known_hosts_file)))
-		goto fail;
-	PathCchConvertStyleA(certificate_store->file, strlen(certificate_store->file), PATH_STYLE_UNIX);
-
-	if (!ensure_path_exists(ConfigPath) || !ensure_path_exists(certificate_store->certs_path) ||
-	    !ensure_path_exists(certificate_store->server_path))
-		goto fail;
-
+	Stream_Write_UINT32(s, blob->length);
+	Stream_Write(s, blob->data, blob->length);
 	return TRUE;
-fail:
-	WLog_ERR(TAG, "certificate store initialization failed");
-	certificate_store_uninit(certificate_store);
+}
+
+void cert_blob_free(rdpCertBlob* blob)
+{
+	if (!blob)
+		return;
+	free(blob->data);
+	blob->data = NULL;
+	blob->length = 0;
+}
+
+/**
+ * Read X.509 Certificate
+ */
+
+static BOOL certificate_read_x509_certificate(rdpCertBlob* cert, rdpCertInfo* info)
+{
+	wStream sbuffer = { 0 };
+	wStream* s = NULL;
+	size_t length = 0;
+	BYTE padding = 0;
+	UINT32 version = 0;
+	size_t modulus_length = 0;
+	size_t exponent_length = 0;
+	int error = 0;
+
+	if (!cert || !info)
+		return FALSE;
+
+	cert_info_free(info);
+	s = Stream_StaticConstInit(&sbuffer, cert->data, cert->length);
+
+	if (!s)
+		return FALSE;
+
+	if (!ber_read_sequence_tag(s, &length)) /* Certificate (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	if (!ber_read_sequence_tag(s, &length)) /* TBSCertificate (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	if (!ber_read_contextual_tag(s, 0, &length, TRUE)) /* Explicit Contextual Tag [0] */
+		goto error;
+
+	error++;
+
+	if (!ber_read_integer(s, &version)) /* version (INTEGER) */
+		goto error;
+
+	error++;
+	version++;
+
+	/* serialNumber */
+	if (!ber_read_integer(s, NULL)) /* CertificateSerialNumber (INTEGER) */
+		goto error;
+
+	error++;
+
+	/* signature */
+	if (!ber_read_sequence_tag(s, &length) ||
+	    !Stream_SafeSeek(s, length)) /* AlgorithmIdentifier (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	/* issuer */
+	if (!ber_read_sequence_tag(s, &length) || !Stream_SafeSeek(s, length)) /* Name (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	/* validity */
+	if (!ber_read_sequence_tag(s, &length) || !Stream_SafeSeek(s, length)) /* Validity (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	/* subject */
+	if (!ber_read_sequence_tag(s, &length) || !Stream_SafeSeek(s, length)) /* Name (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	/* subjectPublicKeyInfo */
+	if (!ber_read_sequence_tag(s, &length)) /* SubjectPublicKeyInfo (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	/* subjectPublicKeyInfo::AlgorithmIdentifier */
+	if (!ber_read_sequence_tag(s, &length) ||
+	    !Stream_SafeSeek(s, length)) /* AlgorithmIdentifier (SEQUENCE) */
+		goto error;
+
+	error++;
+
+	/* subjectPublicKeyInfo::subjectPublicKey */
+	if (!ber_read_bit_string(s, &length, &padding)) /* BIT_STRING */
+		goto error;
+
+	error++;
+
+	/* RSAPublicKey (SEQUENCE) */
+	if (!ber_read_sequence_tag(s, &length)) /* SEQUENCE */
+		goto error;
+
+	error++;
+
+	if (!ber_read_integer_length(s, &modulus_length)) /* modulus (INTEGER) */
+		goto error;
+
+	error++;
+
+	/* skip zero padding, if any */
+	do
+	{
+		if (!Stream_CheckAndLogRequiredLength(TAG, s, 1))
+			goto error;
+
+		Stream_Peek_UINT8(s, padding);
+
+		if (padding == 0)
+		{
+			if (!Stream_SafeSeek(s, 1))
+				goto error;
+
+			modulus_length--;
+		}
+	} while (padding == 0);
+
+	error++;
+
+	if (!cert_info_read_modulus(info, modulus_length, s))
+		goto error;
+
+	error++;
+
+	if (!ber_read_integer_length(s, &exponent_length)) /* publicExponent (INTEGER) */
+		goto error;
+
+	error++;
+
+	if (!cert_info_read_exponent(info, exponent_length, s))
+		goto error;
+	return TRUE;
+error:
+	WLog_ERR(TAG, "error reading when reading certificate: part=%s error=%d",
+	         certificate_read_errors[error], error);
+	cert_info_free(info);
 	return FALSE;
 }
 
-static int compare_pem(const char* current, const char* stored)
+/**
+ * Instantiate new X.509 Certificate Chain.
+ * @param count certificate chain count
+ * @return new X.509 certificate chain
+ */
+
+static rdpX509CertChain certificate_new_x509_certificate_chain(UINT32 count)
 {
-	int rc = 1;
-	X509* xcur = NULL;
-	X509* xstore = NULL;
-	char* fpcur = NULL;
-	char* fpstore = NULL;
-	if (!current || !stored)
-		goto fail;
+	rdpX509CertChain x509_cert_chain = { 0 };
 
-	xcur = crypto_cert_from_pem(current, strlen(current), FALSE);
-	xstore = crypto_cert_from_pem(stored, strlen(stored), FALSE);
-	if (!xcur || !xstore)
-		goto fail;
+	x509_cert_chain.array = (rdpCertBlob*)calloc(count, sizeof(rdpCertBlob));
 
-	fpcur = crypto_cert_fingerprint(xcur);
-	fpstore = crypto_cert_fingerprint(xstore);
-	if (!fpcur || !fpstore)
-		goto fail;
+	if (x509_cert_chain.array)
+		x509_cert_chain.count = count;
 
-	rc = strcmp(fpcur, fpstore);
-
-fail:
-	free(fpcur);
-	free(fpstore);
-	X509_free(xcur);
-	X509_free(xstore);
-	return rc;
+	return x509_cert_chain;
 }
 
-static int certificate_data_match_raw(rdpCertificateStore* certificate_store,
-                                      const rdpCertificateData* certificate_data, char** psubject,
-                                      char** pissuer, char** fprint, char** ppem)
+/**
+ * Free X.509 Certificate Chain.
+ * @param x509_cert_chain X.509 certificate chain to be freed
+ */
+
+static void certificate_free_x509_certificate_chain(rdpX509CertChain* x509_cert_chain)
 {
-	BOOL found = FALSE;
-	HANDLE fp;
-	char* data = NULL;
-	char* mdata = NULL;
-	char* pline = NULL;
-	int match = 1;
-	DWORD lowSize, highSize;
-	UINT64 size;
+	if (!x509_cert_chain)
+		return;
 
-	DWORD read;
-
-	fp = open_file(certificate_store->file, GENERIC_READ, FILE_SHARE_READ, OPEN_ALWAYS,
-	               FILE_ATTRIBUTE_HIDDEN | FILE_ATTRIBUTE_NORMAL);
-
-	if (fp == INVALID_HANDLE_VALUE)
-		return match;
-
-	if ((lowSize = GetFileSize(fp, &highSize)) == INVALID_FILE_SIZE)
+	if (x509_cert_chain->array)
 	{
-		WLog_ERR(TAG, "GetFileSize(%s) returned %s [0x%08" PRIX32 "]", certificate_store->file,
-		         strerror(errno), GetLastError());
-		goto fail;
-	}
-
-	size = (UINT64)lowSize | ((UINT64)highSize << 32);
-
-	if (size < 1)
-		goto fail;
-
-	mdata = (char*)malloc(size + 2);
-
-	if (!mdata)
-		goto fail;
-
-	data = mdata;
-
-	if (!ReadFile(fp, data, size, &read, NULL) || (read != size))
-		goto fail;
-
-	data[size] = '\n';
-	data[size + 1] = '\0';
-	pline = StrSep(&data, "\r\n");
-
-	while ((pline != NULL) && !found)
-	{
-		size_t length = strlen(pline);
-		rdpCertificateData* cert = NULL;
-
-		if (length == 0)
-			goto next;
-
-		if (certificate_line_is_comment(pline, length))
-			goto next;
-
-		cert = certificate_split_line(pline);
-		if (!cert)
+		for (UINT32 i = 0; i < x509_cert_chain->count; i++)
 		{
-			WLog_WARN(TAG, "Invalid %s entry %s!", certificate_known_hosts_file, pline);
-			goto next;
+			rdpCertBlob* element = &x509_cert_chain->array[i];
+			cert_blob_free(element);
 		}
-
-		{
-			const char* old = certificate_data_get_host(cert);
-			const char* cur = certificate_data_get_host(certificate_data);
-
-			if (strcmp(old, cur) != 0)
-				goto next;
-		}
-
-		if (certificate_data_get_port(cert) != certificate_data_get_port(certificate_data))
-			goto next;
-
-		{
-			const char* fingerprint = certificate_data_get_fingerprint(cert);
-			const char* subject = certificate_data_get_subject(cert);
-			const char* issuer = certificate_data_get_issuer(cert);
-			const char* pem = certificate_data_get_pem(cert);
-			const char* cur_fp = certificate_data_get_fingerprint(certificate_data);
-			const char* cur_pem = certificate_data_get_pem(certificate_data);
-
-			duplicate(psubject, subject);
-			duplicate(pissuer, issuer);
-			duplicate(fprint, fingerprint);
-			duplicate(ppem, pem);
-
-			match = -1;
-			if (fingerprint && cur_fp)
-				match = (strcmp(cur_fp, fingerprint) == 0) ? 0 : -1;
-
-			if (cur_pem && pem)
-				match = compare_pem(cur_pem, pem);
-
-			found = TRUE;
-		}
-
-	next:
-		certificate_data_free(cert);
-		pline = StrSep(&data, "\r\n");
 	}
 
-fail:
-	free(mdata);
-	CloseHandle(fp);
-
-	return match;
+	free(x509_cert_chain->array);
 }
 
-static WCHAR* certificate_get_cert_file_name(rdpCertificateStore* store,
-                                             const rdpCertificateData* data)
+static BOOL update_x509_from_info(rdpCertificate* cert)
 {
-	size_t x, offset = 0;
-	char* pem = NULL;
-	WCHAR* wpem = NULL;
-	WINPR_DIGEST_CTX* ctx = NULL;
-	BYTE hash[WINPR_SHA3_256_DIGEST_LENGTH] = { 0 };
-	char fname[WINPR_SHA3_256_DIGEST_LENGTH * 2 + 6] = { 0 };
-
-	if (!store || !store->server_path || !data || !data->hostname)
-		goto fail;
-
-	ctx = winpr_Digest_New();
-	if (!ctx)
-		goto fail;
-	if (!winpr_Digest_Init(ctx, WINPR_MD_SHA256))
-		goto fail;
-	if (!winpr_Digest_Update(ctx, (const BYTE*)data->hostname, strlen(data->hostname)))
-		goto fail;
-	if (!winpr_Digest_Update(ctx, (const BYTE*)&data->port, sizeof(data->port)))
-		goto fail;
-	if (!winpr_Digest_Final(ctx, hash, sizeof(hash)))
-		goto fail;
-
-	for (x = 0; x < sizeof(hash); x++)
-	{
-		int rc = _snprintf(&fname[offset], sizeof(fname) - offset, "%02x", hash[x]);
-		if (rc != 2)
-			goto fail;
-		offset += (size_t)rc;
-	}
-	_snprintf(&fname[offset], sizeof(fname) - offset, ".pem");
-	pem = GetCombinedPath(store->server_path, fname);
-	if (!pem)
-		goto fail;
-
-	wpem = ConvertUtf8ToWCharAlloc(pem, NULL);
-fail:
-	free(pem);
-	winpr_Digest_Free(ctx);
-	return wpem;
-}
-
-static BOOL duplicate(char** data, const char* value)
-{
-	char* tmp = NULL;
-	if (!data)
-		return FALSE;
-	if (value)
-		tmp = _strdup(value);
-
-	free(*data);
-	*data = tmp;
-	if ((value == tmp) && (tmp == NULL))
-		return TRUE;
-	return tmp != NULL;
-}
-
-static rdpCertificateData* load_from_file(rdpCertificateStore* store, const char* hostname,
-                                          UINT16 port)
-{
-	rdpCertificateData* loaded;
-	if (!store || !hostname)
-		return NULL;
-
-	loaded = certificate_data_new(hostname, port);
-	if (!loaded)
-		return NULL;
-	if (!certificate_get_file_data(store, loaded))
-	{
-		certificate_data_free(loaded);
-		return NULL;
-	}
-	return loaded;
-}
-
-static BOOL certificate_get_file_data(rdpCertificateStore* store, rdpCertificateData* data)
-{
-	DWORD read;
 	BOOL rc = FALSE;
-	HANDLE fp = INVALID_HANDLE_VALUE;
 
-	X509* certificate = NULL;
-	LARGE_INTEGER position = { 0 };
-	const LARGE_INTEGER offset = { 0 };
-	WCHAR* fname = certificate_get_cert_file_name(store, data);
-	if (!fname)
+	WINPR_ASSERT(cert);
+
+	X509_free(cert->x509);
+	cert->x509 = NULL;
+
+	rdpCertInfo* info = &cert->cert_info;
+
+	RSA* rsa = RSA_new();
+	BIGNUM* e = BN_new();
+	BIGNUM* mod = BN_new();
+	if (!mod || !e || !rsa)
+		goto fail;
+	if (!BN_bin2bn(info->Modulus, info->ModulusLength, mod))
+		goto fail;
+	if (!BN_bin2bn(info->exponent, sizeof(info->exponent), e))
 		goto fail;
 
-	fp = CreateFileW(fname, GENERIC_READ, 0, NULL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL);
-	if (fp == INVALID_HANDLE_VALUE)
+	const int rec = RSA_set0_key(rsa, mod, e, NULL);
+	if (rec != 1)
 		goto fail;
 
-	if (!SetFilePointerEx(fp, offset, &position, FILE_END))
-		goto fail;
-	if (!SetFilePointerEx(fp, offset, NULL, FILE_BEGIN))
-		goto fail;
-
-	{
-		char* pem = realloc(data->pem, position.QuadPart + 1);
-		if (!pem)
-			goto fail;
-
-		data->pem = pem;
-	}
-
-	if (!ReadFile(fp, data->pem, position.QuadPart, &read, NULL))
-		goto fail;
-	data->pem[read] = '\0';
-
-	certificate = crypto_cert_from_pem(data->pem, read, FALSE);
-	if (!certificate)
-		goto fail;
-
-	free(data->fingerprint);
-	data->fingerprint = crypto_cert_fingerprint(certificate);
-	if (!data->fingerprint)
-		goto fail;
-	free(data->issuer);
-	data->issuer = crypto_cert_issuer(certificate);
-	if (!data->issuer)
-		goto fail;
-	free(data->subject);
-	data->subject = crypto_cert_subject(certificate);
-	if (!data->subject)
+	cert->x509 = x509_from_rsa(rsa);
+	if (!cert->x509)
 		goto fail;
 
 	rc = TRUE;
 
 fail:
-	CloseHandle(fp);
-	X509_free(certificate);
-	free(fname);
-	return rc;
-}
-
-static int certificate_match_data_file(rdpCertificateStore* certificate_store,
-                                       const rdpCertificateData* certificate_data)
-{
-	int rc = 1;
-	const char* pem;
-	const char* loaded_pem;
-
-	rdpCertificateData* loaded =
-	    load_from_file(certificate_store, certificate_data->hostname, certificate_data->port);
-	if (!loaded)
-		return 1;
-
-	pem = certificate_data_get_pem(certificate_data);
-	loaded_pem = certificate_data_get_pem(loaded);
-	if (!pem)
-	{
-		const char* fp = certificate_data_get_fingerprint(certificate_data);
-		const char* load_fp = certificate_data_get_fingerprint(loaded);
-		if (fp)
-		{
-			if (strcmp(fp, load_fp) == 0)
-				rc = 0;
-			else
-				rc = -1;
-		}
-	}
+	if (rsa)
+		RSA_free(rsa);
 	else
 	{
-		rc = compare_pem(pem, loaded_pem);
+		BN_free(mod);
+		BN_free(e);
 	}
-
-	certificate_data_free(loaded);
+	cert->isRSA = rc;
 	return rc;
 }
 
-static BOOL useKnownHosts(rdpCertificateStore* certificate_store)
+static BOOL certificate_process_server_public_key(rdpCertificate* cert, wStream* s, UINT32 length)
 {
-	BOOL use;
-	WINPR_ASSERT(certificate_store);
+	char magic[sizeof(rsa_magic)] = { 0 };
+	UINT32 keylen = 0;
+	UINT32 bitlen = 0;
+	UINT32 datalen = 0;
+	BYTE* tmp = NULL;
 
-	use = freerdp_settings_get_bool(certificate_store->settings, FreeRDP_CertificateUseKnownHosts);
-	WLog_DBG(TAG, "known_hosts=%d", use);
-	return use;
-}
+	WINPR_ASSERT(cert);
+	WINPR_ASSERT(s);
 
-#define check_certificate_store_and_data(store, data) \
-	check_certificate_store_and_data_((store), (data), __FILE__, __FUNCTION__, __LINE__)
-static BOOL check_certificate_store_and_data_(const rdpCertificateStore* certificate_store,
-                                              const rdpCertificateData* certificate_data,
-                                              const char* file, const char* fkt, size_t line)
-{
-	if (!certificate_store)
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, 20))
+		return FALSE;
+
+	Stream_Read(s, magic, sizeof(magic));
+
+	if (memcmp(magic, rsa_magic, sizeof(magic)) != 0)
 	{
-		WLog_ERR(TAG, "[%s, %s:%" PRIuz "] certificate_store=NULL", file, fkt, line);
+		WLog_ERR(TAG, "magic error");
 		return FALSE;
 	}
-	if (!certificate_data)
+
+	cert->isRSA = TRUE;
+	rdpCertInfo* info = &cert->cert_info;
+	cert_info_free(info);
+
+	Stream_Read_UINT32(s, keylen);
+	Stream_Read_UINT32(s, bitlen);
+	Stream_Read_UINT32(s, datalen);
+	Stream_Read(s, info->exponent, 4);
+
+	if ((keylen <= 8) || (!Stream_CheckAndLogRequiredLength(TAG, s, keylen)))
+		return FALSE;
+
+	info->ModulusLength = keylen - 8;
+	tmp = realloc(info->Modulus, info->ModulusLength);
+
+	if (!tmp)
+		return FALSE;
+	info->Modulus = tmp;
+
+	Stream_Read(s, info->Modulus, info->ModulusLength);
+	Stream_Seek(s, 8); /* 8 bytes of zero padding */
+	return update_x509_from_info(cert);
+}
+
+static BOOL certificate_process_server_public_signature(rdpCertificate* certificate,
+                                                        const BYTE* sigdata, size_t sigdatalen,
+                                                        wStream* s, UINT32 siglen)
+{
+	WINPR_ASSERT(certificate);
+#if defined(CERT_VALIDATE_PADDING) || defined(CERT_VALIDATE_RSA)
+	size_t i, sum;
+#endif
+#if defined(CERT_VALIDATE_RSA)
+	BYTE sig[TSSK_KEY_LENGTH];
+#endif
+	BYTE encsig[TSSK_KEY_LENGTH + 8];
+#if defined(CERT_VALIDATE_MD5) && defined(CERT_VALIDATE_RSA)
+	BYTE md5hash[WINPR_MD5_DIGEST_LENGTH];
+#endif
+#if !defined(CERT_VALIDATE_MD5) || !defined(CERT_VALIDATE_RSA)
+	(void)sigdata;
+	(void)sigdatalen;
+#endif
+	(void)certificate;
+	/* Do not bother with validation of server proprietary certificate. The use of MD5 here is not
+	 * allowed under FIPS. Since the validation is not protecting against anything since the
+	 * private/public keys are well known and documented in MS-RDPBCGR section 5.3.3.1, we are not
+	 * gaining any security by using MD5 for signature comparison. Rather then use MD5
+	 * here we just dont do the validation to avoid its use. Historically, freerdp has been ignoring
+	 * a failed validation anyways. */
+#if defined(CERT_VALIDATE_MD5)
+
+	if (!winpr_Digest(WINPR_MD_MD5, sigdata, sigdatalen, md5hash, sizeof(md5hash)))
+		return FALSE;
+
+#endif
+	Stream_Read(s, encsig, siglen);
+
+	if (siglen < 8)
+		return FALSE;
+
+		/* Last 8 bytes shall be all zero. */
+#if defined(CERT_VALIDATE_PADDING)
+
+	for (sum = 0, i = sizeof(encsig) - 8; i < sizeof(encsig); i++)
+		sum += encsig[i];
+
+	if (sum != 0)
 	{
-		WLog_ERR(TAG, "[%s, %s:%" PRIuz "] certificate_data=NULL", file, fkt, line);
+		WLog_ERR(TAG, "invalid signature");
+		return FALSE;
+	}
+
+#endif
+#if defined(CERT_VALIDATE_RSA)
+
+	if (crypto_rsa_public_decrypt(encsig, siglen - 8, TSSK_KEY_LENGTH, tssk_modulus, tssk_exponent,
+	                              sig) <= 0)
+	{
+		WLog_ERR(TAG, "invalid RSA decrypt");
+		return FALSE;
+	}
+
+	/* Verify signature. */
+	/* Do not bother with validation of server proprietary certificate as described above. */
+#if defined(CERT_VALIDATE_MD5)
+
+	if (memcmp(md5hash, sig, sizeof(md5hash)) != 0)
+	{
+		WLog_ERR(TAG, "invalid signature");
+		return FALSE;
+	}
+
+#endif
+	/*
+	 * Verify rest of decrypted data:
+	 * The 17th byte is 0x00.
+	 * The 18th through 62nd bytes are each 0xFF.
+	 * The 63rd byte is 0x01.
+	 */
+
+	for (sum = 0, i = 17; i < 62; i++)
+		sum += sig[i];
+
+	if (sig[16] != 0x00 || sum != 0xFF * (62 - 17) || sig[62] != 0x01)
+	{
+		WLog_ERR(TAG, "invalid signature");
+		return FALSE;
+	}
+
+#endif
+	return TRUE;
+}
+
+static BOOL certificate_read_server_proprietary_certificate(rdpCertificate* certificate, wStream* s)
+{
+	UINT32 dwSigAlgId = 0;
+	UINT32 dwKeyAlgId = 0;
+	UINT16 wPublicKeyBlobType = 0;
+	UINT16 wPublicKeyBlobLen = 0;
+	UINT16 wSignatureBlobType = 0;
+	UINT16 wSignatureBlobLen = 0;
+	BYTE* sigdata = NULL;
+	size_t sigdatalen = 0;
+
+	WINPR_ASSERT(certificate);
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, 12))
+		return FALSE;
+
+	/* -4, because we need to include dwVersion */
+	sigdata = Stream_Pointer(s) - 4;
+	Stream_Read_UINT32(s, dwSigAlgId);
+	Stream_Read_UINT32(s, dwKeyAlgId);
+
+	if (!((dwSigAlgId == SIGNATURE_ALG_RSA) && (dwKeyAlgId == KEY_EXCHANGE_ALG_RSA)))
+	{
+		WLog_ERR(TAG,
+		         "unsupported signature or key algorithm, dwSigAlgId=%" PRIu32
+		         " dwKeyAlgId=%" PRIu32 "",
+		         dwSigAlgId, dwKeyAlgId);
+		return FALSE;
+	}
+
+	Stream_Read_UINT16(s, wPublicKeyBlobType);
+
+	if (wPublicKeyBlobType != BB_RSA_KEY_BLOB)
+	{
+		WLog_ERR(TAG, "unsupported public key blob type %" PRIu16 "", wPublicKeyBlobType);
+		return FALSE;
+	}
+
+	Stream_Read_UINT16(s, wPublicKeyBlobLen);
+
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, wPublicKeyBlobLen))
+		return FALSE;
+
+	if (!certificate_process_server_public_key(certificate, s, wPublicKeyBlobLen))
+		return FALSE;
+
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, 4))
+		return FALSE;
+
+	sigdatalen = Stream_Pointer(s) - sigdata;
+	Stream_Read_UINT16(s, wSignatureBlobType);
+
+	if (wSignatureBlobType != BB_RSA_SIGNATURE_BLOB)
+	{
+		WLog_ERR(TAG, "unsupported blob signature %" PRIu16 "", wSignatureBlobType);
+		return FALSE;
+	}
+
+	Stream_Read_UINT16(s, wSignatureBlobLen);
+
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, wSignatureBlobLen))
+		return FALSE;
+
+	if (wSignatureBlobLen != 72)
+	{
+		WLog_ERR(TAG, "invalid signature length (got %" PRIu16 ", expected 72)", wSignatureBlobLen);
+		return FALSE;
+	}
+
+	if (!certificate_process_server_public_signature(certificate, sigdata, sigdatalen, s,
+	                                                 wSignatureBlobLen))
+	{
+		WLog_ERR(TAG, "unable to parse server public signature");
 		return FALSE;
 	}
 	return TRUE;
 }
 
-int certificate_store_contains_data(rdpCertificateStore* certificate_store,
-                                    const rdpCertificateData* certificate_data)
+/* [MS-RDPBCGR] 2.2.1.4.3.1.1.1 RSA Public Key (RSA_PUBLIC_KEY) */
+static BOOL cert_write_rsa_public_key(wStream* s, const rdpCertificate* cert)
 {
-	if (!check_certificate_store_and_data(certificate_store, certificate_data))
+	WINPR_ASSERT(cert);
+	WINPR_ASSERT(cert->isRSA);
+
+	const rdpCertInfo* info = &cert->cert_info;
+
+	const UINT32 keyLen = info->ModulusLength + 8;
+	const UINT32 bitLen = info->ModulusLength * 8;
+	const UINT32 dataLen = (bitLen / 8) - 1;
+	const size_t pubExpLen = sizeof(info->exponent);
+	const BYTE* pubExp = info->exponent;
+	const BYTE* modulus = info->Modulus;
+
+	const UINT16 wPublicKeyBlobLen = 16 + pubExpLen + keyLen;
+	if (!Stream_EnsureRemainingCapacity(s, 2 + wPublicKeyBlobLen))
+		return FALSE;
+	Stream_Write_UINT16(s, wPublicKeyBlobLen);
+	Stream_Write(s, rsa_magic, sizeof(rsa_magic));
+	Stream_Write_UINT32(s, keyLen);
+	Stream_Write_UINT32(s, bitLen);
+	Stream_Write_UINT32(s, dataLen);
+	Stream_Write(s, pubExp, pubExpLen);
+	Stream_Write(s, modulus, info->ModulusLength);
+	Stream_Zero(s, 8);
+	return TRUE;
+}
+
+static BOOL cert_write_rsa_signature(wStream* s, const void* sigData, size_t sigDataLen)
+{
+	BYTE encryptedSignature[TSSK_KEY_LENGTH] = { 0 };
+	BYTE signature[sizeof(initial_signature)] = { 0 };
+
+	memcpy(signature, initial_signature, sizeof(initial_signature));
+	if (!winpr_Digest(WINPR_MD_MD5, sigData, sigDataLen, signature, sizeof(signature)))
+		return FALSE;
+
+	crypto_rsa_private_encrypt(signature, sizeof(signature), priv_key_tssk, encryptedSignature,
+	                           sizeof(encryptedSignature));
+
+	if (!Stream_EnsureRemainingCapacity(s, 2 * sizeof(UINT16) + sizeof(encryptedSignature) + 8))
+		return FALSE;
+	Stream_Write_UINT16(s, BB_RSA_SIGNATURE_BLOB);
+	Stream_Write_UINT16(s, sizeof(encryptedSignature) + 8); /* wSignatureBlobLen */
+	Stream_Write(s, encryptedSignature, sizeof(encryptedSignature));
+	Stream_Zero(s, 8);
+	return TRUE;
+}
+
+/* [MS-RDPBCGR] 2.2.1.4.3.1.1 Server Proprietary Certificate (PROPRIETARYSERVERCERTIFICATE) */
+static BOOL cert_write_server_certificate_v1(wStream* s, const rdpCertificate* certificate)
+{
+	const size_t start = Stream_GetPosition(s);
+	const BYTE* sigData = Stream_Pointer(s) - sizeof(UINT32);
+
+	WINPR_ASSERT(start >= 4);
+	if (!Stream_EnsureRemainingCapacity(s, 10))
+		return FALSE;
+	Stream_Write_UINT32(s, SIGNATURE_ALG_RSA);
+	Stream_Write_UINT32(s, KEY_EXCHANGE_ALG_RSA);
+	Stream_Write_UINT16(s, BB_RSA_KEY_BLOB);
+	if (!cert_write_rsa_public_key(s, certificate))
+		return FALSE;
+
+	const size_t end = Stream_GetPosition(s);
+	return cert_write_rsa_signature(s, sigData, end - start + sizeof(UINT32));
+}
+
+static BOOL cert_write_server_certificate_v2(wStream* s, const rdpCertificate* certificate)
+{
+	WINPR_ASSERT(certificate);
+
+	const rdpX509CertChain* chain = &certificate->x509_cert_chain;
+	const size_t padding = 8ull + 4ull * chain->count;
+
+	if (Stream_EnsureRemainingCapacity(s, sizeof(UINT32)))
+		return FALSE;
+
+	Stream_Write_UINT32(s, chain->count);
+	for (UINT32 x = 0; x < chain->count; x++)
+	{
+		const rdpCertBlob* cert = &chain->array[x];
+		if (!cert_blob_write(cert, s))
+			return FALSE;
+	}
+
+	if (Stream_EnsureRemainingCapacity(s, padding))
+		return FALSE;
+	Stream_Zero(s, padding);
+	return FALSE;
+}
+
+SSIZE_T freerdp_certificate_write_server_cert(const rdpCertificate* certificate, UINT32 dwVersion,
+                                              wStream* s)
+{
+	if (!certificate)
 		return -1;
-	if (useKnownHosts(certificate_store))
+
+	const size_t start = Stream_GetPosition(s);
+	if (!Stream_EnsureRemainingCapacity(s, 4))
+		return -1;
+	Stream_Write_UINT32(s, dwVersion);
+
+	switch (dwVersion & CERT_CHAIN_VERSION_MASK)
 	{
-		char* pem = NULL;
-		int rc =
-		    certificate_data_match_raw(certificate_store, certificate_data, NULL, NULL, NULL, &pem);
-
-		/* Upgrade entry, append PEM */
-		if ((rc == 0) && !pem && certificate_data->pem)
-		{
-			certificate_store_save_data(certificate_store, certificate_data);
-		}
-		free(pem);
-		return rc;
+		case CERT_CHAIN_VERSION_1:
+			if (!cert_write_server_certificate_v1(s, certificate))
+				return -1;
+			break;
+		case CERT_CHAIN_VERSION_2:
+			if (!cert_write_server_certificate_v2(s, certificate))
+				return -1;
+			break;
+		default:
+			WLog_ERR(TAG, "invalid certificate chain version:%" PRIu32 "",
+			         dwVersion & CERT_CHAIN_VERSION_MASK);
+			return -1;
 	}
-	else
-		return certificate_match_data_file(certificate_store, certificate_data);
+
+	const size_t end = Stream_GetPosition(s);
+	return end - start;
 }
 
-static char* decode(const char* value)
+/**
+ * Read an X.509 Certificate Chain.
+ * @param certificate certificate module
+ * @param s stream
+ */
+
+static BOOL certificate_read_server_x509_certificate_chain(rdpCertificate* cert, wStream* s)
 {
-	size_t len, length;
-	char* converted = NULL;
-	if (!value)
-		return NULL;
-	len = strlen(value);
-	crypto_base64_decode(value, len, (BYTE**)&converted, &length);
-	return converted;
-}
+	BOOL ret = FALSE;
+	UINT32 numCertBlobs = 0;
+	DEBUG_CERTIFICATE("Server X.509 Certificate Chain");
 
-static char* encode(const char* value)
-{
-	size_t len;
-	if (!value)
-		return NULL;
-	len = strlen(value);
-	return (char*)crypto_base64_encode((const BYTE*)value, len);
-}
-
-static char* allocated_printf(const char* fmt, ...)
-{
-	int rc1, rc2;
-	size_t size;
-	va_list ap;
-	char* buffer = NULL;
-
-	va_start(ap, fmt);
-	rc1 = vsnprintf(NULL, 0, fmt, ap);
-	va_end(ap);
-	if (rc1 <= 0)
-		return NULL;
-
-	size = (size_t)rc1;
-	buffer = calloc(size + 2, sizeof(char));
-	if (!buffer)
-		return NULL;
-
-	va_start(ap, fmt);
-	rc2 = vsnprintf(buffer, size + 1, fmt, ap);
-	va_end(ap);
-	if (rc2 != rc1)
-	{
-		free(buffer);
-		return NULL;
-	}
-	return buffer;
-}
-
-static char* certificate_data_get_host_file_entry(const rdpCertificateData* data)
-{
-	char* buffer = NULL;
-	const char* hostname = certificate_data_get_host(data);
-	const UINT16 port = certificate_data_get_port(data);
-	char* subject = encode(certificate_data_get_subject(data));
-	char* issuer = encode(certificate_data_get_issuer(data));
-	const char* fingerprint = certificate_data_get_fingerprint(data);
-	char* pem = encode(certificate_data_get_pem(data));
-
-	/* Issuer and subject may be NULL */
-	if (!data || !hostname || !fingerprint)
-		goto fail;
-
-	if (!subject)
-		subject = _strdup("");
-	if (!issuer)
-		issuer = _strdup("");
-
-	if (pem)
-		buffer = allocated_printf("%s %" PRIu16 " %s %s %s %s\n", hostname, port, fingerprint,
-		                          subject, issuer, pem);
-	else
-		buffer = allocated_printf("%s %" PRIu16 " %s %s %s\n", hostname, port, fingerprint, subject,
-		                          issuer);
-fail:
-	free(subject);
-	free(issuer);
-	free(pem);
-	return buffer;
-}
-
-static BOOL write_line_and_free(const char* filename, HANDLE fp, char* line)
-{
-	BOOL rc = FALSE;
-	DWORD size, written;
-	if ((fp == INVALID_HANDLE_VALUE) || !line)
+	WINPR_ASSERT(cert);
+	if (!Stream_CheckAndLogRequiredLength(TAG, s, 4))
 		return FALSE;
-	size = strlen(line);
-	rc = WriteFile(fp, line, size, &written, NULL);
 
-	if (!rc || (written != size))
+	Stream_Read_UINT32(s, numCertBlobs); /* numCertBlobs */
+	cert->x509_cert_chain = certificate_new_x509_certificate_chain(numCertBlobs);
+
+	for (UINT32 i = 0; i < cert->x509_cert_chain.count; i++)
 	{
-		WLog_ERR(TAG, "WriteFile(%s) returned %s [0x%08X]", filename, strerror(errno), errno);
-		rc = FALSE;
-	}
+		rdpCertBlob* blob = &cert->x509_cert_chain.array[i];
+		if (!cert_blob_read(blob, s))
+			return FALSE;
 
-	free(line);
-	return rc;
-}
-
-static BOOL certificate_data_replace_hosts_file(rdpCertificateStore* certificate_store,
-                                                const rdpCertificateData* certificate_data,
-                                                BOOL remove, BOOL append)
-{
-	BOOL newfile = TRUE;
-	HANDLE fp = INVALID_HANDLE_VALUE;
-	BOOL rc = FALSE;
-	size_t length;
-	char* data = NULL;
-	char* sdata = NULL;
-	char* pline = NULL;
-
-	UINT64 size;
-	DWORD read;
-	DWORD lowSize, highSize;
-	rdpCertificateData* copy = NULL;
-
-	fp = open_file(certificate_store->file, GENERIC_READ | GENERIC_WRITE, 0, OPEN_ALWAYS,
-	               FILE_ATTRIBUTE_NORMAL);
-
-	if (fp == INVALID_HANDLE_VALUE)
-	{
-		WLog_ERR(TAG, "failed to open known_hosts file %s, aborting", certificate_store->file);
-		return FALSE;
-	}
-
-	/* Create a copy, if a PEM was provided it will replace subject, issuer, fingerprint */
-	copy = certificate_data_new(certificate_data->hostname, certificate_data->port);
-	if (!copy)
-		goto fail;
-
-	if (certificate_data->pem)
-	{
-		if (!certificate_data_set_pem(copy, certificate_data->pem))
-			goto fail;
-	}
-	else
-	{
-		if (!certificate_data_set_subject(copy, certificate_data->subject) ||
-		    !certificate_data_set_issuer(copy, certificate_data->issuer) ||
-		    !certificate_data_set_fingerprint(copy, certificate_data->fingerprint))
-			goto fail;
-	}
-
-	if ((lowSize = GetFileSize(fp, &highSize)) == INVALID_FILE_SIZE)
-	{
-		WLog_ERR(TAG, "GetFileSize(%s) returned %s [0x%08" PRIX32 "]", certificate_store->file,
-		         strerror(errno), GetLastError());
-		goto fail;
-	}
-
-	size = (UINT64)lowSize | ((UINT64)highSize << 32);
-
-	/* Newly created file, just write the new entry */
-	if (size > 0)
-		newfile = FALSE;
-
-	if (newfile)
-		goto fail;
-
-	data = (char*)malloc(size + 2);
-
-	if (!data)
-		goto fail;
-
-	if (!ReadFile(fp, data, size, &read, NULL) || (read != size))
-		goto fail;
-
-	if (SetFilePointer(fp, 0, NULL, FILE_BEGIN) == INVALID_SET_FILE_POINTER)
-	{
-		WLog_ERR(TAG, "SetFilePointer(%s) returned %s [0x%08" PRIX32 "]", certificate_store->file,
-		         strerror(errno), GetLastError());
-		goto fail;
-	}
-
-	if (!SetEndOfFile(fp))
-	{
-		WLog_ERR(TAG, "SetEndOfFile(%s) returned %s [0x%08" PRIX32 "]", certificate_store->file,
-		         strerror(errno), GetLastError());
-		goto fail;
-	}
-
-	/* Write the file back out, with appropriate fingerprint substitutions */
-	data[size] = '\n';
-	data[size + 1] = '\0';
-	sdata = data;
-	pline = StrSep(&sdata, "\r\n");
-
-	while (pline != NULL)
-	{
-		rdpCertificateData* cert = NULL;
-		length = strlen(pline);
-
-		if (length > 0)
+		if ((numCertBlobs - i) == 2)
 		{
-			if (certificate_line_is_comment(pline, length))
+			rdpCertInfo cert_info = { 0 };
+			DEBUG_CERTIFICATE("License Server Certificate");
+			ret = certificate_read_x509_certificate(blob, &cert_info);
+			DEBUG_CERTIFICATE("modulus length:%" PRIu32 "", cert_info.ModulusLength);
+			cert_info_free(&cert_info);
+
+			if (!ret)
 			{
-			}
-			else if (!(cert = certificate_split_line(pline)))
-				WLog_WARN(TAG, "Skipping invalid %s entry %s!", certificate_known_hosts_file,
-				          pline);
-			else
-			{
-				char* line;
-
-				/* If this is the replaced hostname, use the updated fingerprint. */
-				if ((strcmp(certificate_data_get_host(cert), certificate_data_get_host(copy)) ==
-				     0) &&
-				    (certificate_data_get_port(cert) == certificate_data_get_port(copy)))
-				{
-					rc = TRUE;
-					if (remove && !append)
-						goto next;
-
-					line = certificate_data_get_host_file_entry(copy);
-				}
-				else
-					line = certificate_data_get_host_file_entry(cert);
-
-				if (!line)
-					goto next;
-
-				if (!write_line_and_free(certificate_store->file, fp, line))
-					goto fail;
+				WLog_ERR(TAG, "failed to read License Server, content follows:");
+				winpr_HexDump(TAG, WLOG_ERROR, blob->data, blob->length);
+				return FALSE;
 			}
 		}
-	next:
-		certificate_data_free(cert);
-		pline = StrSep(&sdata, "\r\n");
-	}
-
-fail:
-	if (!rc && append)
-	{
-		if (fp != INVALID_HANDLE_VALUE)
+		else if (numCertBlobs - i == 1)
 		{
-			char* line = certificate_data_get_host_file_entry(copy);
-			if (line)
-				rc = write_line_and_free(certificate_store->file, fp, line);
+			DEBUG_CERTIFICATE("Terminal Server Certificate");
+
+			if (!certificate_read_x509_certificate(blob, &cert->cert_info))
+				return FALSE;
+
+			DEBUG_CERTIFICATE("modulus length:%" PRIu32 "", cert->cert_info.ModulusLength);
 		}
 	}
 
-	CloseHandle(fp);
-	free(data);
-	certificate_data_free(copy);
-	return rc;
+	return update_x509_from_info(cert);
 }
 
-static BOOL certificate_data_write_to_file(rdpCertificateStore* certificate_store,
-                                           const rdpCertificateData* certificate_data)
+static BOOL certificate_write_server_x509_certificate_chain(const rdpCertificate* certificate,
+                                                            wStream* s)
 {
-	BOOL rc = FALSE;
-	size_t len;
-	HANDLE handle;
-	DWORD numberOfBytesWritten;
-	WCHAR* fname = certificate_get_cert_file_name(certificate_store, certificate_data);
-	if (!fname)
-		return FALSE;
+	UINT32 numCertBlobs = 0;
 
-	handle = CreateFileW(fname, GENERIC_WRITE, 0, NULL, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL);
-	if (handle == INVALID_HANDLE_VALUE)
-		goto fail;
-	len = strlen(certificate_data->pem);
-	rc = WriteFile(handle, certificate_data->pem, len, &numberOfBytesWritten, NULL);
-	CloseHandle(handle);
-fail:
-	free(fname);
-	return rc;
-}
+	WINPR_ASSERT(certificate);
+	WINPR_ASSERT(s);
 
-static BOOL certificate_data_remove_file(rdpCertificateStore* certificate_store,
-                                         const rdpCertificateData* certificate_data)
-{
-	BOOL rc = FALSE;
-	WCHAR* fname = certificate_get_cert_file_name(certificate_store, certificate_data);
-	if (!fname)
-		return FALSE;
-	if (PathFileExistsW(fname))
-		rc = DeleteFileW(fname);
-	else
-		rc = TRUE;
-	free(fname);
-	return rc;
-}
+	numCertBlobs = certificate->x509_cert_chain.count;
 
-BOOL certificate_store_remove_data(rdpCertificateStore* certificate_store,
-                                   const rdpCertificateData* certificate_data)
-{
-	if (!check_certificate_store_and_data(certificate_store, certificate_data))
+	if (!Stream_EnsureRemainingCapacity(s, 4))
 		return FALSE;
-	if (useKnownHosts(certificate_store))
+	Stream_Write_UINT32(s, numCertBlobs); /* numCertBlobs */
+
+	for (UINT32 i = 0; i < numCertBlobs; i++)
 	{
-		/* Ignore return, if the entry was invalid just continue */
-		certificate_data_replace_hosts_file(certificate_store, certificate_data, TRUE, FALSE);
+		const rdpCertBlob* cert = &certificate->x509_cert_chain.array[i];
+		if (!cert_blob_write(cert, s))
+			return FALSE;
+	}
+
+	return TRUE;
+}
+
+/**
+ * Read a Server Certificate.
+ * @param certificate certificate module
+ * @param server_cert server certificate
+ * @param length certificate length
+ */
+
+BOOL freerdp_certificate_read_server_cert(rdpCertificate* certificate, const BYTE* server_cert,
+                                          size_t length)
+{
+	BOOL ret = FALSE;
+	wStream *s, sbuffer;
+	UINT32 dwVersion = 0;
+
+	WINPR_ASSERT(certificate);
+	if (length < 4) /* NULL certificate is not an error see #1795 */
 		return TRUE;
+
+	WINPR_ASSERT(server_cert);
+	s = Stream_StaticConstInit(&sbuffer, server_cert, length);
+
+	if (!s)
+	{
+		WLog_ERR(TAG, "Stream_New failed!");
+		return FALSE;
 	}
-	else
-		return certificate_data_remove_file(certificate_store, certificate_data);
+
+	Stream_Read_UINT32(s, dwVersion); /* dwVersion (4 bytes) */
+
+	switch (dwVersion & CERT_CHAIN_VERSION_MASK)
+	{
+		case CERT_CHAIN_VERSION_1:
+			ret = certificate_read_server_proprietary_certificate(certificate, s);
+			break;
+
+		case CERT_CHAIN_VERSION_2:
+			ret = certificate_read_server_x509_certificate_chain(certificate, s);
+			break;
+
+		default:
+			WLog_ERR(TAG, "invalid certificate chain version:%" PRIu32 "",
+			         dwVersion & CERT_CHAIN_VERSION_MASK);
+			ret = FALSE;
+			break;
+	}
+
+	return ret;
 }
 
-rdpCertificateData* certificate_split_line(char* line)
+static BOOL cert_blob_copy(rdpCertBlob* dst, const rdpCertBlob* src)
 {
-	char* cur;
-	char* host = NULL;
-	char* subject = NULL;
-	char* issuer = NULL;
-	char* fingerprint = NULL;
-	char* pem = NULL;
-	UINT16 port = 0;
-	rdpCertificateData* data = NULL;
+	const rdpCertBlob empty = { 0 };
+	WINPR_ASSERT(dst);
+	WINPR_ASSERT(src);
 
-	size_t length = strlen(line);
-
-	if (length <= 0)
-		goto fail;
-
-	cur = StrSep(&line, " \t");
-
-	if (!cur)
-		goto fail;
-
-	host = cur;
-	cur = StrSep(&line, " \t");
-
-	if (!cur)
-		goto fail;
-
-	if (sscanf(cur, "%hu", &port) != 1)
-		goto fail;
-
-	cur = StrSep(&line, " \t");
-
-	if (!cur)
-		goto fail;
-
-	fingerprint = cur;
-	cur = StrSep(&line, " \t");
-
-	if (!cur)
-		goto fail;
-
-	subject = cur;
-	cur = StrSep(&line, " \t");
-
-	if (!cur)
-		goto fail;
-
-	issuer = cur;
-
-	/* Optional field */
-	cur = StrSep(&line, " \t");
-	if (cur)
-		pem = cur;
-
-	data = certificate_data_new(host, port);
-	if (!data)
-		goto fail;
-	if (pem)
+	*dst = empty;
+	if (src->length > 0)
 	{
-		BOOL rc;
-		char* dpem = NULL;
-		size_t clength;
-		crypto_base64_decode(pem, strlen(pem), (BYTE**)&dpem, &clength);
-		rc = certificate_data_set_pem(data, dpem);
-		free(dpem);
-		if (!rc)
-			goto fail;
+		dst->data = malloc(src->length);
+		if (!dst->data)
+			return FALSE;
+		dst->length = src->length;
+		memcpy(dst->data, src->data, src->length);
 	}
-	else
-	{
-		BOOL rc;
-		size_t clength;
-		char* dsubject = NULL;
-		char* dissuer = NULL;
-		crypto_base64_decode(subject, strlen(subject), (BYTE**)&dsubject, &clength);
-		crypto_base64_decode(issuer, strlen(issuer), (BYTE**)&dissuer, &clength);
 
-		rc = certificate_data_set_subject(data, dsubject) &&
-		     certificate_data_set_issuer(data, dissuer) &&
-		     certificate_data_set_fingerprint(data, fingerprint);
-		free(dsubject);
-		free(dissuer);
-		if (!rc)
-			goto fail;
+	return TRUE;
+}
+
+static BOOL cert_x509_chain_copy(rdpX509CertChain* cert, const rdpX509CertChain* src)
+{
+	rdpX509CertChain empty = { 0 };
+
+	WINPR_ASSERT(cert);
+
+	*cert = empty;
+	if (!src)
+		return TRUE;
+
+	if (src->count > 0)
+	{
+		cert->array = calloc(src->count, sizeof(rdpCertBlob));
+		if (!cert->array)
+		{
+			return FALSE;
+		}
+		cert->count = src->count;
+
+		for (UINT32 x = 0; x < cert->count; x++)
+		{
+			const rdpCertBlob* srcblob = &src->array[x];
+			rdpCertBlob* dstblob = &cert->array[x];
+
+			if (!cert_blob_copy(dstblob, srcblob))
+			{
+				certificate_free_x509_certificate_chain(cert);
+				return FALSE;
+			}
+		}
 	}
-	return data;
-fail:
-	certificate_data_free(data);
+
+	return TRUE;
+}
+
+BOOL cert_clone_int(rdpCertificate* dst, const rdpCertificate* src)
+{
+	WINPR_ASSERT(dst);
+	WINPR_ASSERT(src);
+
+	dst->isRSA = src->isRSA;
+	if (src->x509)
+	{
+		dst->x509 = X509_dup(src->x509);
+		if (!dst->x509)
+			return FALSE;
+	}
+
+	if (!cert_info_clone(&dst->cert_info, &src->cert_info))
+		return FALSE;
+	return cert_x509_chain_copy(&dst->x509_cert_chain, &src->x509_cert_chain);
+}
+
+rdpCertificate* freerdp_certificate_clone(const rdpCertificate* certificate)
+{
+	if (!certificate)
+		return NULL;
+
+	rdpCertificate* _certificate = freerdp_certificate_new();
+
+	if (!_certificate)
+		return NULL;
+
+	if (!cert_clone_int(_certificate, certificate))
+		goto out_fail;
+
+	return _certificate;
+out_fail:
+
+	freerdp_certificate_free(_certificate);
 	return NULL;
 }
 
-static BOOL update_from_pem(rdpCertificateData* data)
+/**
+ * Instantiate new certificate module.
+ * @return new certificate module
+ */
+
+rdpCertificate* freerdp_certificate_new(void)
+{
+	return (rdpCertificate*)calloc(1, sizeof(rdpCertificate));
+}
+
+void certificate_free_int(rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+
+	if (cert->x509)
+		X509_free(cert->x509);
+	if (cert->chain)
+		sk_X509_free(cert->chain);
+
+	certificate_free_x509_certificate_chain(&cert->x509_cert_chain);
+	cert_info_free(&cert->cert_info);
+}
+
+/**
+ * Free certificate module.
+ * @param certificate certificate module to be freed
+ */
+
+void freerdp_certificate_free(rdpCertificate* cert)
+{
+	if (!cert)
+		return;
+
+	certificate_free_int(cert);
+	free(cert);
+}
+
+static BOOL freerdp_rsa_from_x509(rdpCertificate* cert)
 {
 	BOOL rc = FALSE;
-	char* subject = NULL;
-	char* issuer = NULL;
-	char* fingerprint = NULL;
-	X509* x1 = NULL;
 
-	if (!data || !data->pem)
-		return FALSE;
-	x1 = crypto_cert_from_pem(data->pem, strlen(data->pem), FALSE);
-	if (!x1)
+	WINPR_ASSERT(cert);
+
+	cert->isRSA = FALSE;
+
+	RSA* rsa = NULL;
+	EVP_PKEY* pubkey = X509_get_pubkey(cert->x509);
+	if (!pubkey)
+		goto fail;
+	rsa = EVP_PKEY_get1_RSA(pubkey);
+
+	/* If this is not a RSA key return success */
+	rc = TRUE;
+	if (!rsa)
 		goto fail;
 
-	/* Subject and issuer might be NULL */
-	subject = crypto_cert_subject(x1);
-	issuer = crypto_cert_issuer(x1);
+	/* Now we return failure again if something is wrong. */
+	rc = FALSE;
 
-	fingerprint = crypto_cert_fingerprint(x1);
-	if (!fingerprint)
+	const BIGNUM* rsa_n = NULL;
+	const BIGNUM* rsa_e = NULL;
+	RSA_get0_key(rsa, &rsa_n, &rsa_e, NULL);
+	if (!rsa_n || !rsa_e)
 		goto fail;
-	duplicate(&data->subject, subject);
-	duplicate(&data->issuer, issuer);
-	duplicate(&data->fingerprint, fingerprint);
+	if (!cert_info_create(&cert->cert_info, rsa_n, rsa_e))
+		goto fail;
 	rc = TRUE;
 fail:
-	free(subject);
-	free(issuer);
-	free(fingerprint);
-	X509_free(x1);
+	RSA_free(rsa);
+	EVP_PKEY_free(pubkey);
+	cert->isRSA = rc;
 	return rc;
 }
 
-BOOL certificate_store_save_data(rdpCertificateStore* certificate_store,
-                                 const rdpCertificateData* certificate_data)
+rdpCertificate* freerdp_certificate_new_from_der(const BYTE* data, size_t length)
 {
-	if (!check_certificate_store_and_data(certificate_store, certificate_data))
-		return FALSE;
-	if (useKnownHosts(certificate_store))
-		return certificate_data_replace_hosts_file(certificate_store, certificate_data, TRUE, TRUE);
-	else
-		return certificate_data_write_to_file(certificate_store, certificate_data);
-}
+	rdpCertificate* cert = freerdp_certificate_new();
 
-rdpCertificateData* certificate_store_load_data(rdpCertificateStore* certificate_store,
-                                                const char* host, UINT16 port)
-{
-	if (useKnownHosts(certificate_store))
-	{
-		int rc;
-		rdpCertificateData* data = certificate_data_new(host, port);
-		if (!data)
-			return NULL;
-
-		rc = certificate_data_match_raw(certificate_store, data, &data->subject, &data->issuer,
-		                                &data->fingerprint, &data->pem);
-		if ((rc == 0) || (rc == -1))
-			return data;
-		certificate_data_free(data);
-		return NULL;
-	}
-	else
-	{
-		return load_from_file(certificate_store, host, port);
-	}
-}
-
-rdpCertificateData* certificate_data_new(const char* hostname, UINT16 port)
-{
-	size_t i;
-	rdpCertificateData* certdata = NULL;
-
-	if (!hostname)
+	if (!cert || !data || (length == 0))
 		goto fail;
-
-	certdata = (rdpCertificateData*)calloc(1, sizeof(rdpCertificateData));
-
-	if (!certdata)
+	cert->x509 = d2i_X509(NULL, &data, length);
+	if (!cert->x509)
 		goto fail;
-
-	certdata->port = port;
-	certdata->hostname = _strdup(hostname);
-	if (!certdata->hostname)
+	if (!freerdp_rsa_from_x509(cert))
 		goto fail;
-	for (i = 0; i < strlen(hostname); i++)
-		certdata->hostname[i] = tolower(certdata->hostname[i]);
-
-	return certdata;
+	return cert;
 fail:
-	certificate_data_free(certdata);
+	freerdp_certificate_free(cert);
 	return NULL;
 }
 
-void certificate_data_free(rdpCertificateData* certificate_data)
+rdpCertificate* freerdp_certificate_new_from_x509(const X509* xcert, const STACK_OF(X509) * chain)
 {
-	if (certificate_data != NULL)
+	WINPR_ASSERT(xcert);
+
+	rdpCertificate* cert = freerdp_certificate_new();
+	if (!cert)
+		return NULL;
+
+	cert->x509 = X509_dup(xcert);
+	if (!cert->x509)
+		goto fail;
+
+	if (!freerdp_rsa_from_x509(cert))
+		goto fail;
+
+	if (chain)
 	{
-		free(certificate_data->hostname);
-		free(certificate_data->subject);
-		free(certificate_data->issuer);
-		free(certificate_data->fingerprint);
-		free(certificate_data->pem);
-		free(certificate_data);
+		cert->chain = sk_X509_dup(chain);
 	}
+	return cert;
+fail:
+	freerdp_certificate_free(cert);
+	return NULL;
 }
 
-const char* certificate_data_get_host(const rdpCertificateData* cert)
+static rdpCertificate* freerdp_certificate_new_from(const char* file, BOOL isFile)
 {
-	if (!cert)
+	X509* x509 = x509_utils_from_pem(file, strlen(file), isFile);
+	if (!x509)
 		return NULL;
-	return cert->hostname;
+	rdpCertificate* cert = freerdp_certificate_new_from_x509(x509, NULL);
+	X509_free(x509);
+	return cert;
 }
 
-UINT16 certificate_data_get_port(const rdpCertificateData* cert)
+rdpCertificate* freerdp_certificate_new_from_file(const char* file)
 {
-	if (!cert)
-		return 0;
-	return cert->port;
+	return freerdp_certificate_new_from(file, TRUE);
 }
 
-BOOL certificate_data_set_pem(rdpCertificateData* cert, const char* pem)
+rdpCertificate* freerdp_certificate_new_from_pem(const char* pem)
 {
-	if (!cert)
+	return freerdp_certificate_new_from(pem, FALSE);
+}
+
+const rdpCertInfo* freerdp_certificate_get_info(const rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	if (!cert->isRSA)
+		return NULL;
+	return &cert->cert_info;
+}
+
+char* freerdp_certificate_get_fingerprint(const rdpCertificate* cert)
+{
+	return freerdp_certificate_get_fingerprint_by_hash(cert, "sha256");
+}
+
+char* freerdp_certificate_get_fingerprint_by_hash(const rdpCertificate* cert, const char* hash)
+{
+	return freerdp_certificate_get_fingerprint_by_hash_ex(cert, hash, TRUE);
+}
+
+char* freerdp_certificate_get_fingerprint_by_hash_ex(const rdpCertificate* cert, const char* hash,
+                                                     BOOL separator)
+{
+	size_t fp_len = 0;
+	size_t pos = 0;
+	size_t size = 0;
+	BYTE* fp = NULL;
+	char* fp_buffer = NULL;
+	if (!cert || !cert->x509)
 	{
-		return FALSE;
-	}
-	if (!duplicate(&cert->pem, pem))
-		return FALSE;
-	if (!pem)
-		return TRUE;
-
-	return update_from_pem(cert);
-}
-
-BOOL certificate_data_set_subject(rdpCertificateData* cert, const char* subject)
-{
-	if (!cert)
-		return FALSE;
-	return duplicate(&cert->subject, subject);
-}
-BOOL certificate_data_set_issuer(rdpCertificateData* cert, const char* issuer)
-{
-	if (!cert)
-		return FALSE;
-	return duplicate(&cert->issuer, issuer);
-}
-BOOL certificate_data_set_fingerprint(rdpCertificateData* cert, const char* fingerprint)
-{
-	if (!cert)
-		return FALSE;
-	return duplicate(&cert->fingerprint, fingerprint);
-}
-const char* certificate_data_get_pem(const rdpCertificateData* cert)
-{
-	if (!cert)
-		return NULL;
-	return cert->pem;
-}
-
-const char* certificate_data_get_subject(const rdpCertificateData* cert)
-{
-	if (!cert)
-		return NULL;
-	return cert->subject;
-}
-const char* certificate_data_get_issuer(const rdpCertificateData* cert)
-{
-	if (!cert)
-		return NULL;
-	return cert->issuer;
-}
-const char* certificate_data_get_fingerprint(const rdpCertificateData* cert)
-{
-	if (!cert)
-		return NULL;
-	return cert->fingerprint;
-}
-
-rdpCertificateStore* certificate_store_new(const rdpSettings* settings)
-{
-	rdpCertificateStore* certificate_store =
-	    (rdpCertificateStore*)calloc(1, sizeof(rdpCertificateStore));
-
-	if (!certificate_store)
-		return NULL;
-
-	certificate_store->settings = settings;
-
-	if (!certificate_store_init(certificate_store))
-	{
-		certificate_store_free(certificate_store);
+		WLog_ERR(TAG, "Invalid certificate [%p, %p]", cert, cert ? cert->x509 : NULL);
 		return NULL;
 	}
-
-	return certificate_store;
-}
-
-void certificate_store_free(rdpCertificateStore* certstore)
-{
-	certificate_store_uninit(certstore);
-	free(certstore);
-}
-
-const char* certificate_store_get_hosts_file(const rdpCertificateStore* certificate_store)
-{
-	if (!certificate_store)
+	if (!hash)
+	{
+		WLog_ERR(TAG, "Invalid certificate hash %p", hash);
 		return NULL;
-	return certificate_store->file;
+	}
+	fp = x509_utils_get_hash(cert->x509, hash, &fp_len);
+	if (!fp)
+		return NULL;
+
+	size = fp_len * 3 + 1;
+	fp_buffer = calloc(size, sizeof(char));
+	if (!fp_buffer)
+		goto fail;
+
+	pos = 0;
+
+	size_t i = 0;
+	for (i = 0; i < (fp_len - 1); i++)
+	{
+		int rc;
+		char* p = &fp_buffer[pos];
+		if (separator)
+			rc = sprintf_s(p, size - pos, "%02" PRIx8 ":", fp[i]);
+		else
+			rc = sprintf_s(p, size - pos, "%02" PRIx8, fp[i]);
+		if (rc <= 0)
+			goto fail;
+		pos += (size_t)rc;
+	}
+
+	sprintf_s(&fp_buffer[pos], size - pos, "%02" PRIx8 "", fp[i]);
+
+	free(fp);
+
+	return fp_buffer;
+fail:
+	free(fp);
+	free(fp_buffer);
+	return NULL;
 }
 
-const char* certificate_store_get_certs_path(const rdpCertificateStore* certificate_store)
+static BOOL bio_read_pem(BIO* bio, char** ppem, size_t* plength)
 {
-	if (!certificate_store)
-		return NULL;
-	return certificate_store->certs_path;
+	BOOL rc = FALSE;
+
+	WINPR_ASSERT(bio);
+	WINPR_ASSERT(ppem);
+
+	size_t offset = 0;
+	size_t length = 2048;
+	char* pem = NULL;
+	while (offset < length)
+	{
+		char* tmp = realloc(pem, length + 1);
+		if (!tmp)
+			goto fail;
+		pem = tmp;
+
+		ERR_clear_error();
+
+		const int status = BIO_read(bio, &pem[offset], length - offset);
+		if (status < 0)
+		{
+			WLog_ERR(TAG, "failed to read certificate");
+			goto fail;
+		}
+
+		if (status == 0)
+			break;
+
+		offset += (size_t)status;
+		if (length - offset > 0)
+			break;
+		length *= 2;
+	}
+	pem[offset] = '\0';
+	*ppem = pem;
+	if (plength)
+		*plength = offset;
+	rc = TRUE;
+fail:
+	return rc;
 }
 
-const char* certificate_store_get_hosts_path(const rdpCertificateStore* certificate_store)
+char* freerdp_certificate_get_pem(const rdpCertificate* cert, size_t* pLength)
 {
+	char* pem = NULL;
+	WINPR_ASSERT(cert);
 
-	if (!certificate_store)
+	if (!cert->x509)
 		return NULL;
-	return certificate_store->server_path;
+
+	BIO* bio;
+	int status;
+	BOOL rc = FALSE;
+
+	/**
+	 * Don't manage certificates internally, leave it up entirely to the external client
+	 * implementation
+	 */
+	bio = BIO_new(BIO_s_mem());
+
+	if (!bio)
+	{
+		WLog_ERR(TAG, "BIO_new() failure");
+		return NULL;
+	}
+
+	status = PEM_write_bio_X509(bio, cert->x509);
+
+	if (status < 0)
+	{
+		WLog_ERR(TAG, "PEM_write_bio_X509 failure: %d", status);
+		goto fail;
+	}
+
+#if 0
+        if (chain)
+        {
+            count = sk_X509_num(chain);
+            for (x = 0; x < count; x++)
+            {
+                X509* c = sk_X509_value(chain, x);
+                status = PEM_write_bio_X509(bio, c);
+                if (status < 0)
+                {
+                    WLog_ERR(TAG, "PEM_write_bio_X509 failure: %d", status);
+                    goto fail;
+                }
+            }
+        }
+#endif
+	if (!bio_read_pem(bio, &pem, pLength))
+		goto fail;
+	rc = TRUE;
+fail:
+	BIO_free_all(bio);
+	return pem;
+}
+
+char* freerdp_certificate_get_subject(const rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_subject(cert->x509);
+}
+
+char* freerdp_certificate_get_issuer(const rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_issuer(cert->x509);
+}
+
+char* freerdp_certificate_get_upn(const rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_upn(cert->x509);
+}
+
+char* freerdp_certificate_get_email(const rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_email(cert->x509);
+}
+
+BOOL freerdp_certificate_check_eku(const rdpCertificate* cert, int nid)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_check_eku(cert->x509, nid);
+}
+
+BOOL freerdp_certificate_get_public_key(const rdpCertificate* cert, BYTE** PublicKey,
+                                        DWORD* PublicKeyLength)
+{
+	BYTE* ptr;
+	int length;
+	BOOL status = TRUE;
+	EVP_PKEY* pkey = NULL;
+
+	WINPR_ASSERT(cert);
+
+	pkey = X509_get_pubkey(cert->x509);
+
+	if (!pkey)
+	{
+		WLog_ERR(TAG, "X509_get_pubkey() failed");
+		status = FALSE;
+		goto exit;
+	}
+
+	length = i2d_PublicKey(pkey, NULL);
+
+	if (length < 1)
+	{
+		WLog_ERR(TAG, "i2d_PublicKey() failed");
+		status = FALSE;
+		goto exit;
+	}
+
+	*PublicKeyLength = (DWORD)length;
+	*PublicKey = (BYTE*)malloc(length);
+	ptr = (BYTE*)(*PublicKey);
+
+	if (!ptr)
+	{
+		status = FALSE;
+		goto exit;
+	}
+
+	i2d_PublicKey(pkey, &ptr);
+exit:
+
+	if (pkey)
+		EVP_PKEY_free(pkey);
+
+	return status;
+}
+
+BOOL freerdp_certificate_verify(const rdpCertificate* cert, const char* certificate_store_path)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_verify(cert->x509, cert->chain, certificate_store_path);
+}
+
+char** freerdp_certificate_get_dns_names(const rdpCertificate* cert, size_t* pcount,
+                                         size_t** pplengths)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_dns_names(cert->x509, pcount, pplengths);
+}
+
+char* freerdp_certificate_get_common_name(const rdpCertificate* cert, size_t* plength)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_common_name(cert->x509, plength);
+}
+
+WINPR_MD_TYPE freerdp_certificate_get_signature_alg(const rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_signature_alg(cert->x509);
+}
+
+void freerdp_certificate_free_dns_names(size_t count, size_t* lengths, char** names)
+{
+	x509_utils_dns_names_free(count, lengths, names);
+}
+
+char* freerdp_certificate_get_hash(const rdpCertificate* cert, const char* hash, size_t* plength)
+{
+	WINPR_ASSERT(cert);
+	return x509_utils_get_hash(cert->x509, hash, plength);
+}
+
+X509* freerdp_certificate_get_x509(rdpCertificate* cert)
+{
+	WINPR_ASSERT(cert);
+	return cert->x509;
+}
+
+BYTE* freerdp_certificate_get_der(const rdpCertificate* cert, size_t* pLength)
+{
+	WINPR_ASSERT(cert);
+
+	if (pLength)
+		*pLength = 0;
+
+	BYTE* ptr = NULL;
+	const int rc = i2d_X509(cert->x509, &ptr);
+	if (rc < 0)
+		return NULL;
+
+	if (pLength)
+		*pLength = (size_t)rc;
+	return ptr;
 }
